requirements.txt
tensorflow
torch
torchvision
scikit-learn
opencv-python
numpy
pandas
matplotlib
streamlit
fastapi
uvicorn
Pillow
